requirements.txt
anyio==3.6.2
black==23.3.0
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
coverage==7.2.6
cryptography==40.0.2
docutils==0.20.1
dotty-dict==1.3.1
ecdsa==0.18.0
exceptiongroup==1.1.1
execnet==1.9.0
fastapi==0.95.2
fastapi_class==3.3.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
invoke==1.7.3
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.13.1
more-itertools==9.1.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pytest==7.3.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
python-dotenv==1.0.0
python-gitlab==3.14.0
python-jose==3.3.0
python-semantic-release==7.33.5
PyYAML==6.0
readme-renderer==37.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rsa==4.9
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
zipp==3.15.0
